Changed various file names + updated Homework FSM
Renamed FSM files for more clarity between homework and PAMI FSM
Added Backward and Stop states and deleted turnleft and turnright function
