--- robot/motors/Motor.h
@@ -1,10 +1,0 @@
-#ifndef MOTOR_H
-#define MOTOR_H
-
-class Motor {
-public:
-    void init();
-    void setSpeed(int speed);
-};
-
-#endif
