--- robot/sensors/Ultrason.h
@@ -1,11 +1,0 @@
-// Ultrason.h
-#ifndef ULTRASON_H
-#define ULTRASON_H
-
-class Ultrason {
-public:
-    void init();
-    int readDistance();
-};
-
-#endif
